GRPC SERVER READY

+++ internal/models/user.go
@@ -1,11 +1,10 @@
 package models
 
 type User struct {
-	ID        int
+	ID        int64
 	Login     string
 	FirstName string
 	LastName  string
 	Email     string
 	Phone     string
 }
-
